+++ src/AST/Visitor.php
@@ -6,9 +6,11 @@ interface Visitor
 {
     public function visitAndExpression(AndExpression $node);
     public function visitElement(Element $node);
+    public function visitElementList(ElementList $node);
     public function visitExpression(Expression $node);
     public function visitIdExpression(IdExpression $node);
     public function visitOpExpression(OpExpression $node);
     public function visitValue(Value $node);
     public function visitValueList(ValueList $node);
+    public function visitFunc(Func $node);
 }