+++ test/visitor.php
@@ -2,11 +2,6 @@
 
 use Krak\AQL;
 
-beforeEach(function() {
-    $this->parser = new AQL\Parser\ExpressionParser();
-    $this->compiler = new AQL\Compiler\ExpressionCompiler();
-});
-
 describe('DoubleToSingleQuotesVisitor', function() {
     it('transforms double quoted string to single quoted strings', function() {
         $ast = $this->parser->parse('"s"');